Crate fedimint_wallet_common

Modules

config

endpoint_constants

envs

keys

tweakable

txoproof

Structs

PegOut

PegOutFees

PegOutSignatureItem

Rbf

Allows a user to bump the fees of a PendingTransaction

SpendableUTXO

TxOutputSummary

A transaction output, either unspent or consumed

UnknownWalletInputVariantError

UnknownWalletOutputOutcomeVariantError

UnknownWalletOutputVariantError

WalletCommonInit

WalletInputV0

WalletInputV1

WalletModuleTypes

WalletOutputOutcomeV0

Contains the Bitcoin transaction id of the transaction created by the withdraw request

WalletSummary

Summary of the coins within the wallet.

Enums

ProcessPegOutSigError

WalletConsensusItem

WalletInput

WalletInputError

WalletOutput

WalletOutputError

WalletOutputOutcome

WalletOutputV0

Constants

DEPRECATED_RBF_ERROR

FEERATE_MULTIPLIER_DEFAULT

To further mitigate the risk of a peg-out transaction getting stuck in the mempool, we multiply the feerate estimate returned from the backend by this value.

KIND

MODULE_CONSENSUS_VERSION

SAFE_DEPOSIT_MODULE_CONSENSUS_VERSION

Module consensus version that introduced support for processing Bitcoin transactions that exceed the ALEPH_BFT_UNIT_BYTE_LIMIT.

Functions

proprietary_tweak_key

Type Aliases

PartialSig

PegInDescriptor