fedimint_wallet_common

Module config

Structs

FeeConsensus

WalletClientConfig

WalletConfig

WalletConfigConsensus

WalletConfigLocal

WalletConfigPrivate

WalletGenParams

WalletGenParamsConsensus

WalletGenParamsLocal

Constants

DEFAULT_DEPOSIT_FEE_SATS 🔒

Helps against dust attacks where an attacker deposits UTXOs that, with higher fee levels, cannot be spent profitably.