Crate fedimint_wallet_common

Modules

config

db

keys

tweakable

txoproof

Structs

PegOut

PegOutFees

PegOutSignatureItem

PendingTransaction

A peg-out tx that is ready to be broadcast with a tweak for the change UTXO

Rbf

Allows a user to bump the fees of a PendingTransaction

SpendableUTXO

UnsignedTransaction

A PSBT that is awaiting enough signatures from the federation to becoming a PendingTransaction

UnzipWalletConsensusItem

WalletCommonInit

WalletInput

WalletModuleTypes

WalletOutputOutcome

Contains the Bitcoin transaction id of the transaction created by the withdraw request

Enums

ProcessPegOutSigError

WalletConsensusItem

WalletCreationError

WalletInputError

WalletOutput

WalletOutputError

Constants

CONFIRMATION_TARGET

CONSENSUS_VERSION 🔒

KIND

Traits

IterUnzipWalletConsensusItem

Functions

proprietary_tweak_key

Type Definitions

PartialSig

PegInDescriptor