Trait fedimint_wallet_common::tweakable::Contract

pub trait Contract {
    // Required method
    fn encode<W: Write>(&self, writer: &mut W) -> Result<()>;
}

An object that can be used as a Ricardian contract to tweak a key

Required Methods

fn encode<W: Write>(&self, writer: &mut W) -> Result<()>

Serialize the contract in a deterministic way to be used as a tweak

Object Safety

This trait is not object safe.