fedimint_wallet_common::tweakable
Trait Contract

pub trait Contract {
    // Required method
    fn encode<W: Write>(&self, writer: &mut W) -> Result<()>;
}

An object that can be used as a Ricardian contract to tweak a key

Required Methods

fn encode<W: Write>(&self, writer: &mut W) -> Result<()>

Serialize the contract in a deterministic way to be used as a tweak

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.